An unsigned comparison is unsigned, and a uint variable says so once
`uge` decompiled as `>=` between ints. HLSL takes the signedness of a
comparison from its operands and not from the operator, so an int loop
counter turned every ult and uge back into an ilt and an ige. The two agree
on everything below 2^31, which is why no tier ever noticed.

Casting at the comparison was the first version of this and the wrong one:
correct, but `(uint)t2 < 4` on a counter that starts at zero, in eighteen
fixtures. Declaring the variable uint is correct and shorter, and is what
the source said in the first place.

IsUnsignedValue sits beside IsIntegerValue and has the same shape: the
readers where they agree, what made it otherwise. Readers that know are an
ordered integer comparison, the value side of a shift right, a udiv - there
is no signed integer divide opcode, so an integer quotient is unsigned by
construction - and an ineg, which says signed. Readers that carry rather
than judge are a move, a phi, an add, a multiply, a bitwise operator, and a
select's branches but not its condition; imin and umin carry too, since the
parse makes one node of both.

The effect is fewer casts, not more, because a uint variable makes its own
shifts logical and its own comparisons unsigned:

    t0 = ((uint)t0 >> 16) ^ 1664525 * t0 + 1013904223;
    t0 = (t0 >> 16) ^ 1664525 * t0 + 1013904223;

    for (int t0 = 32; t0 > 0; t0 = (uint)t0 >> 1)
    for (uint t0 = 32; t0 > 0; t0 = t0 >> 1)

Where a cast is still needed it goes on the side that is not a constant, and
only when neither side is unsigned already - HLSL promotes the other to
match. Deciding that takes a walk through the promoting operations, since
`1664525 * k.x + 1013904223` over a uint4 k is unsigned arithmetic and
casting into it says nothing. The same test retires three casts the shift
path used to emit over uint registers and thread ids.

skin_buffer's counter is negated and also compared unsigned, so it stays int
and carries its casts. That is the truth about the shader.

The instruction writer had the bug too and keeps the cast form: every temp it
declares is an int4, so it has no variable to retype. Same shape as its ushr.

unsigned_compare is a fixture that can fail. Two earlier attempts could not:
over a uint4 the arithmetic is unsigned by promotion whatever the writer
believes, and an int constant cannot help either, because the interpreter
fills one with `(uint)Math.Abs(value * 2) % 4` on purpose so that a loop
bound read as unsigned still terminates. `k.x * 2654435761u + 7` overflows
into the top bit instead, and turning the fix off fails Equivalent on three
trials.

packed_colour loses a hoist. `(float)((uint)i.color >> 24)` is 28 characters
and `(float)(i.color >> 24)` is 22, against a RepeatedTextBudget of 24, so
what was named is now written twice. Probed rather than assumed: at 16 the
hoist returns and reads better than before, and fourteen more fixtures move
with it, which is its own change.

597 fast, 580 recompile, 288 equivalent (2 known skips), 290 cost. 0 failed.

+++ HlslDecompiler.Tests/ShaderSources/cs_4_0/bitpack.fx
@@ -9,7 +9,7 @@ void main(uint3 sv_dispatchthreadid : SV_DispatchThreadID)
 	if (sv_dispatchthreadid.x >= count) {
 		return;
 	}
-	int t0 = source.Load(sv_dispatchthreadid.x * 4);
-	int2 t1 = int2(t0 & 255, ((uint)77 * ((uint)t0 >> 24) + 151 * (((uint)t0 >> 16) & 255) + 28 * (((uint)t0 >> 8) & 255) >> 8) * 16777216);
-	target.Store(sv_dispatchthreadid.x * 4, (((uint)77 * ((uint)t0 >> 24) + 151 * (((uint)t0 >> 16) & 255) + 28 * (((uint)t0 >> 8) & 255) >> 8) * 65536) + t1.y + (77 * ((uint)t0 >> 24) + 151 * (((uint)t0 >> 16) & 255) + 28 * (((uint)t0 >> 8) & 255) & 65280) + t1.x);
+	uint t0 = source.Load(sv_dispatchthreadid.x * 4);
+	int2 t1 = int2(t0 & 255, (77 * (t0 >> 24) + 151 * ((t0 >> 16) & 255) + 28 * ((t0 >> 8) & 255) >> 8) * 16777216);
+	target.Store(sv_dispatchthreadid.x * 4, ((77 * (t0 >> 24) + 151 * ((t0 >> 16) & 255) + 28 * ((t0 >> 8) & 255) >> 8) * 65536) + t1.y + (77 * (t0 >> 24) + 151 * ((t0 >> 16) & 255) + 28 * ((t0 >> 8) & 255) & 65280) + t1.x);
 }
