+++ HlslDecompiler.Tests/ShaderSources/cs_4_1/groupshared_reduce.fx
@@ -14,7 +14,7 @@ void main(CS_IN i)
 {
 	g0[i.sv_groupindex] = input[(i.sv_groupid.x * 64) + i.sv_groupindex];
 	GroupMemoryBarrierWithGroupSync();
-	for (int t0 = 32; t0 > 0; t0 = (uint)t0 >> 1) {
+	for (uint t0 = 32; t0 > 0; t0 = t0 >> 1) {
 		float t1;
 		if (i.sv_groupindex < t0) {
 			t1 = g0[i.sv_groupindex];
