+++ HlslDecompiler.Tests/ShaderSources/vs_4_0/skin_buffer.fx
@@ -22,7 +22,7 @@ float4 main(VS_IN i) : SV_Position
 	float3 t0 = 0;
 	int t1 = 0;
 	while (t1 < 3) {
-		int t2 = (((t1 < 3 ? -1 : 0) == 0 ? i.blendindices.w : 0) | ((t1 < 2 ? 0 : t1 - 3) & i.blendindices.z)) | (((t1 < 2 ? -t1 : 0) & i.blendindices.y) | ((t1 < 1 ? -1 : 0) & i.blendindices.x));
+		int t2 = ((((uint)t1 < 3 ? -1 : 0) == 0 ? i.blendindices.w : 0) | (((uint)t1 < 2 ? 0 : t1 - 3) & i.blendindices.z)) | ((((uint)t1 < 2 ? -t1 : 0) & i.blendindices.y) | (((uint)t1 < 1 ? -1 : 0) & i.blendindices.x));
 		float t3 = dot(i.blendweight, icb[t1]);
 		t0 = float3(dot(bones.Load(3 * t2), i.position), dot(bones.Load(3 * t2 + 1), i.position), dot(bones.Load(3 * t2 + 2), i.position)) * t3 + t0;
 		t1 = t1 + 1;
